add and update

+++ dbMoved2Data.py
@@ -1,3 +1,4 @@
+#!/usr/bin/python3
 import os
 import shutil
 import eaCLEAN_UP_TEMP
@@ -30,30 +31,32 @@ def __moveSingleFaceData():
 	for x in folder:
 		person=os.listdir("./{0}{1}{2}".format(SS,x,SS))
 		for unit in person:
-			if(__findUnknown(unit)):
-				src=".{0}{1}{2}".format(SS,x,SS)+unit
-				dst=".{0}FR_DATA{1}B-Unknown{2}".format(SS,SS,SS)+unit
-				#print(src+" to "+dst)
-				shutil.move(src,dst)
-			else:
-				src=".{0}{1}{2}".format(SS,x,SS)+unit
-				dst=".{0}FR_DATA{1}D-Singleface{2}".format(SS,SS,SS)+__returnPersonName(unit)+SS+unit
-				personDir = ".{0}FR_DATA{1}D-Singleface{2}".format(SS,SS,SS)+__returnPersonName(unit)
-				if not os.path.exists(personDir):
-					print("Mkdir...{0}".format(__returnPersonName(unit)))
-					os.makedirs(personDir)	# 创建目录
+			if unit != ".keep":
+				if (__findUnknown(unit)):
+					src = ".{0}{1}{2}".format(SS, x, SS) + unit
+					dst = ".{0}FR_DATA{1}B-Unknown{2}".format(SS, SS, SS) + unit
+					# print(src+" to "+dst)
+					shutil.move(src, dst)
 				else:
-					pass
-				shutil.move(src,dst)
+					src = ".{0}{1}{2}".format(SS, x, SS) + unit
+					dst = ".{0}FR_DATA{1}D-Singleface{2}".format(SS, SS, SS) + __returnPersonName(unit) + SS + unit
+					personDir = ".{0}FR_DATA{1}D-Singleface{2}".format(SS, SS, SS) + __returnPersonName(unit)
+					if not os.path.exists(personDir):
+						print("Mkdir...{0}".format(__returnPersonName(unit)))
+						os.makedirs(personDir)  # 创建目录
+					else:
+						pass
+					shutil.move(src, dst)
 
 def __moveMultiFacesData():
 	person=os.listdir(".{0}tempMore{1}".format(SS,SS))
 	#print(person)
 	for unit in person:
-		src=".{0}tempMore{1}".format(SS,SS)+unit
-		dst=".{0}FR_DATA{1}E-Morefaces{2}".format(SS,SS,SS)+unit
-		#print(src+"  to  "+dst)
-		shutil.move(src,dst)
+		if unit != ".keep":
+			src = ".{0}tempMore{1}".format(SS, SS) + unit
+			dst = ".{0}FR_DATA{1}E-Morefaces{2}".format(SS, SS, SS) + unit
+			# print(src+"  to  "+dst)
+			shutil.move(src, dst)
 
 #mainX
 def Moving():
